logging added, thrust defined to 0, squeleton completed.
Print functions have been added to log the data on every
iteration of the loop.

Thrust is defined as the command given by the controller
(teleoperator). Different values should be tested.

The stabilize esqueleton code is fairly complete. Minor
adjustment will be needed but it should be enough for the
first tests. Next steps include a more rigurous implementation
of the dynamical model of the quadrotor and a motor_inversion
module that obtains the motor voltages (u_mi)
from the inputs (Ui).