Updated stabilizer according to IMU interface.
Minor changes in IMU to include a timing parameter that will log
the seconds per measurement. It's not so clear if this is going
to be useful because the sensor timing will be adjusted
according to the control loop leaving (usually) enough time
to the sensor to have new data ready.