Add more motor_inversion methods to dynamical mode.
More inversion methods have been added to the dynamical model module.
The crazyflie (motor_inversion2) replicates the code implemented
in the Crazyflie nano-quadcopter.

After reviewing the code and the behaviour of the robot a problem
seemed clear; the actuation was the oposite than the one expected.
This is because the coordinates of the theoretical models where
opposite to the ones adopted in the robot (both the crazyflie and
M.Wieremma's thesis adopt the same frame references and rotation
directions.)

The easiest fix for this situation is to modify the code to assign
M1 <-> M3 and M2 <-> M4. This has been done in "motor_inversion3"
(taking the crazyflie code as starting point). The result is
satisfactory and it seems that adjusting the PIDs, stable flight might
be achieved.