Motors stop when joystick released.
Motors speed adjusted to respond to the BT controller. If the thrust
desired (z_d) is 0 (either because there's no controller or because
the controller is at 0) the thrust of all motors is forced to 0.

If this hack is not applied the motors tend to move slightly to reach
the position desired.